requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0

# Optional - nếu cần database
# sqlalchemy==2.0.23
# alembic==1.13.1
# asyncpg==0.29.0  # for PostgreSQL
# aiomysql==0.2.0  # for MySQL

# Optional - nếu cần Redis cho caching
# redis==5.0.1
# aioredis==2.0.1

# Optional - nếu cần xử lý images
# Pillow==10.1.0
# opencv-python==4.8.1.78

# Optional - nếu cần AI/NLP
# openai==1.3.0
# langchain==0.0.350
